+++ app/scripts/api.js
@@ -473,13 +473,14 @@ const createApi = function api(context, pubSub) {
        *
        * @param {string} viewUid The identifier of the view to zoom
        * @param {string} geneName The name of gene symbol to search
+       * @param {string} padding The padding (base pairs) around a given gene for the navigation
        * @param {Number} animateTime The time to spend zooming to the specified location
        * @example
        * // Zoom to the location near 'MYC'
-       * hgApi.zoomToGene('view1', 'MYC', 2000);
+       * hgApi.zoomToGene('view1', 'MYC', 100, 2000);
        */
-      zoomToGene(viewUid, geneName, animateTime = 0) {
-        self.zoomToGene(viewUid, geneName, animateTime);
+      zoomToGene(viewUid, geneName, padding = 0, animateTime = 0) {
+        self.zoomToGene(viewUid, geneName, padding, animateTime);
       },
 
       /**