+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Make vertical chromosome labels as well as loading status labels readable in vertical tracks
 - Added an option menu item for rectangle domain fill opacity
+- Added a parameter in `zoomToGene` to allow specifying padding around gene
 - Add data fetchers to `AVAILABLE_FOR_PLUGINS`
 - Update track list in `AVAILABLE_FOR_PLUGINS`
 - Correctly setup initial scales of vertical tracks when the width of a center track is zero.